move stories to feature file
